test: 구분자 생성 팩토리를 두 개로 분리함에 따른 테스트 분리
